[hotfix] Move RetrievableStateStorageHelper and implementation out of zookeeper package so that it could be reused by Kubernetes high-availability
