+++ flink-runtime/src/main/java/org/apache/flink/runtime/persistence/RetrievableStateStorageHelper.java
@@ -16,9 +16,10 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.zookeeper;
+package org.apache.flink.runtime.persistence;
 
 import org.apache.flink.runtime.state.RetrievableStateHandle;
+import org.apache.flink.runtime.zookeeper.ZooKeeperStateHandleStore;
 
 import java.io.Serializable;
